Add a theme selector in the settings
Power to the people!

There is now no theme to select on the app config file. All CSS files
present in the `client/themes` folder will be given as choices to the
users.
Note that this switch will cause existing configurations to be reset to
the default theme.

This is temporary (as in, temporary for a fairly long time) until we
have proper theme management.